Rename pager.ts file
