+++ src/js/components/pager/pager.ts
